requirements.txt
langchain
langchain-google-genai
google-generativeai
flask
python-dotenv